Re-organize the structure of the az and http plugins
This PR re-organize the structure of the az and http plugins
so that it is easy to manage. Now each file system is on its own
directory `tensorflow_io/core/plugins/xxx`. This is for the upcoming
s3/gcs/hdfs plugins migration.
